# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Kanban Backend API - OpenAPI 3.0
  description: API para gerenciamento de projetos Kanban, responsáveis e indicadores.
  contact:
    email: leduar.araujo@hotmail.com
  version: 1.0.0
...
paths:
...
  /secretaria:
    get:
      operationId: listarSecretarias
      tags:
        - Secretaria
      summary: Lista todas as secretarias
      description: ''

      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Secretaria'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'

    post:
      operationId: cadastrarSecretaria
      tags:
        - Secretaria
      summary: Cadastra uma nova secretaria
      description: ''
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nmSecretaria:
                  type: string
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
  /secretaria/{id}:
    get:
      operationId: buscarSecretariaPorId
      tags:
        - Secretaria
      summary: Busca uma secretaria pelo ID
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          description: ID da secretaria a ser buscada
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Secretaria'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    put:
      operationId: atualizarSecretariaPorId
      tags:
        - Secretaria
      summary: Atualiza uma secretaria existente
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          description: ID da secretaria a ser atualizada
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nmSecretaria:
                  type: string
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
    delete:
      operationId: excluirSecretaria
      tags:
        - Secretaria
      summary: Exclui uma secretaria existente
      description: ''
      parameters:
        - name: id
          in: path
          required: true
          description: ID da secretaria a ser excluída
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
...
components:
  schemas:
...
    Secretaria:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nmSecretaria:
          type: string
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: "Sucesso"
        code:
          type: integer
          example: 0
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
          example: "Error message"
        errorCode:
          type: integer
          example: 123
  securitySchemes:
    gerenciador:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write: escrita
            read: leitura

    api_key:
      type: apiKey
      name: api_key
      in: header